requirements.txt
speechrecognition
pyaudio
google-generativeai
